Add parseValue and isColumnOriented options
Adds two options. One to be able to generate a
JSON object based on the columns of the excel
file, and the second to be able to parse values
into integer or booleans when they have such
shape.
Also removes some eslint properties which are n
suported any longer.
Updated readme accordingly.